# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Food-Delivery",
    "description": "<b>RESPONSES</b>::\n 200-> OK\n   401->Not Authorized\n   201->No Record\n 400->Bad Request \n 500->Internal Server Error",
    
      "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
...
   "/api/mobile/orders/getCancelReasons": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/orders/getCancelReasons",
       "operationId": "index",
       "tags": ["Get Cancelation Reason Order"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"}
       
    ],
       "responses": {}
     }
   },
...
   "/api/mobile/orders/list": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/orders/list",
       "operationId": "index",
       "tags": ["Orders Listing"],
       "description": "Orders Listing",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "progressStatus",  "in": "query",  "required":true ,"type":"string"}
       
    ],
       "responses": {}
     }
   },
   "/api/mobile/orders/detail/{id}": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/orders/detail/{id}",
       "operationId": "index",
       "tags": ["Order Info"],
       "description": "Orders Detail",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "id",  "in": "path",  "required":true ,"type":"string"}

    ],
       "responses": {}
     }
   },
...
   "/api/mobile/orders/cancel": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/orders/cancel",
       "operationId": "index",
       "tags": ["Cancel Order"],
       "description": "Orders Cancellation",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "orderId",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "cancellationReason",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "otherReason",  "in": "formData","type":"string"}

    ],
       "responses": {}
     }
   },
   "/api/mobile/orders/status": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/orders/status",
       "operationId": "index",
       "tags": ["Change Order Status"],
       "description": "Order Status  5->Complete",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "id",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "status",  "in": "formData",  "required":true ,"type":"string"}

    ],
       "responses": {}
     }
   },
...
   "/api/mobile/tiffin/orders/list": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/tiffin/orders/list",
       "operationId": "index",
       "tags": ["Get Tiffin Order Listing"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "limit",  "in": "query", "type":"string"},
         { "name": "page",  "in": "query", "type":"string"},
         { "name": "progressStatus",  "in": "query", "type":"string", "description":"e.g : 0,1,2,3,4,5   >>More Info :  0>Pending, 1->Confirmed, 2->Cancelled, 3->Started ,5 ->Completed"}

        ],
       "responses": {}
     }
   },
   "/api/mobile/tiffin/orders/detail/{orderId}": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/tiffin/orders/list",
       "operationId": "index",
       "tags": ["Get Tiffin Order Detail"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "orderId",  "in": "path", "type":"string","required":true}

        ],
       "responses": {}
     }
   },
   "/api/mobile/tiffin/orders/cancel": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/tiffin/orders/cancel",
       "operationId": "index",
       "tags": ["Cancel Tiffin Order"],
       "description": "",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "orderId",  "in": "formData", "type":"string","required":true},
         { "name": "cancellationReason",  "in": "formData", "type":"string"}

        ],
       "responses": {}
     }
   },
...
   "/api/mobile/rest/orders/list": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/rest/orders/list",
       "operationId": "index",
       "tags": ["Restaurant Orders Listing"],
       "description": "Orders Listing",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "progressStatus",  "in": "formData",  "required":true ,"type":"string"}
       
    ],
       "responses": {}
     }
   },
...
   "/api/mobile/rest/orders/allStatus": {
     "get": {
       "x-swagger-router-controller": "/api/mobile/rest/orders/allStatus",
       "operationId": "index",
       "tags": ["Get all Restaurant Status"],
       "description": "Restaurant Order Status ",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"}

    ],
       "responses": {}
     }
   },
   "/api/mobile/rest/orders/status": {
     "post": {
       "x-swagger-router-controller": "/api/mobile/rest/orders/status",
       "operationId": "index",
       "tags": ["Restaurant Change Order Status"],
       "description": "Restaurant Order Status ",
       "parameters": [
         { "name": "authorization",  "in": "header",  "required":true,"type":"string"},
         { "name": "id",  "in": "formData",  "required":true ,"type":"string"},
         { "name": "status",  "in": "formData",  "required":true ,"type":"string"}

    ],
       "responses": {}
     }
   },
...
  }
}